Refactor - remove duplicate DamageCantBePreventedEffect classes
